pyproject.toml
[project]
name = "globalsolutions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-groq>=0.3.2",
    "langchain[anthropic]>=0.3.25",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
